Move logcat code to separate module
